unix socket server
I added an -fpr flag to sf v1.1 (but only when built with archivematica tag). This is a simple server that just implements the FPR interface (it only sends puid + error message) and runs on unix sockets - so faster than the http/json approach.
 
This python script is a client for the -fpr server. I've copied most of Misty's existing script for this with a slight change: on multiple matches this script will return an error rather than giving the first result (I feel this is safer bc when you get multiple matches - which only happens for results that are matched on extension only - they all have equal weight).

Hope this all makes sense!